+++ cmd/coder/main.go
@@ -4,12 +4,11 @@ import (
 	"fmt"
 	"os"
 
-	distcli "github.com/fluxplane/agentruntime/adapters/distribution/cli"
 	"github.com/fluxplane/agentruntime/apps/coder"
 )
 
 func main() {
-	cmd := distcli.NewCommand(coder.Distribution())
+	cmd := coder.NewCommand()
 	if err := cmd.Execute(); err != nil {
 		fmt.Fprintln(os.Stderr, err)
 		os.Exit(1)